Refactor to extract FileUpload logic into hook and new component for adornment
